Add initial fabric training code

+++ src/model/cm/src/nn.py
@@ -127,10 +127,13 @@ def timestep_embedding(timesteps, dim, max_period=10000):
     :return: an [N x dim] Tensor of positional embeddings.
     """
     half = dim // 2
+    # freqs = th.exp(
+    #     -math.log(max_period) * th.arange(start=0, end=half, dtype=th.float32) / half
+    # ).to(device=timesteps.device)
     freqs = th.exp(
-        -math.log(max_period) * th.arange(start=0, end=half, dtype=th.float32) / half
+        -math.log(max_period) * th.arange(start=0, end=half) / half
     ).to(device=timesteps.device)
-    args = timesteps[:, None].float() * freqs[None]
+    args = timesteps[:, None] * freqs[None]
     embedding = th.cat([th.cos(args), th.sin(args)], dim=-1)
     if dim % 2:
         embedding = th.cat([embedding, th.zeros_like(embedding[:, :1])], dim=-1)
